Search Bar integrated into Browse Function
